pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomy-timer"
version = "0.1.0"
description = "Jira time tracking CLI — reads duration_ns from YAML and updates Jira"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.12",
    "requests>=2.31",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.scripts]
autonomy_timer = "jira_tt.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
